# api/openapi.json
openapi: 3.0.0
info:
  title: SlayFlashcards API
  description: REST API for SlayFlashcards - An interactive flashcard learning platform
  version: 1.0.0
servers:
  - url: /api/v1
...
security:
  - bearerAuth: []
paths:
...
  /users:
    get:
      summary: Get all users
      tags:
        - users
      description: Get all users with optional filtering and pagination.
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: limit
          in: query
          description: Items per page
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: name_contains
          in: query
          description: Filter users by name pattern
          schema:
            type: string
        - name: has_sessions
          in: query
          description: Filter users who have sessions
          schema:
            type: boolean
      responses:
        '200':
          description: Retrieved list of users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsersResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Create a new user
      tags:
        - users
      description: Create a new user.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '409':
          description: User with the given name already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/{user_id}:
    get:
      summary: Get a specific user by ID
      tags:
        - users
      description: Get a specific user by ID.
      security:
        - bearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          description: The ID of the user to retrieve.
          schema:
            type: integer
      responses:
        '200':
          description: User retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: Update an existing user
      tags:
        - users
      description: Update an existing user.
      security:
        - bearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          description: The ID of the user to update.
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: User with the new name already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete a user
      tags:
        - users
      description: Delete a user and optionally their sessions.
      security:
        - bearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          description: The ID of the user to delete.
          schema:
            type: integer
        - name: delete_sessions
          in: query
          description: Also delete all user sessions
          schema:
            type: boolean
            default: false
      responses:
        '204':
          description: User deleted successfully
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/name/{username}:
    get:
      summary: Get a user by username
      tags:
        - users
      description: Get a user by username.
      security:
        - bearerAuth: []
      parameters:
        - name: username
          in: path
          required: true
          description: The username of the user to retrieve.
          schema:
            type: string
      responses:
        '200':
          description: User retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User with the given name not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/ensure:
    post:
      summary: Get user if exists, create if not
      tags:
        - users
      description: Get user if exists, create if not.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '200':
          description: User retrieved or created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/{user_id}/statistics:
    get:
      summary: Get comprehensive statistics for a user
      tags:
        - users
      description: Get comprehensive statistics for a user.
      security:
        - bearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          description: The ID of the user to get statistics for.
          schema:
            type: integer
      responses:
        '200':
          description: User statistics retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserStatsResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/{user_id}/progress:
    get:
      summary: Get user learning progress over time
      tags:
        - users
      description: Get user learning progress over time.
      security:
        - bearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          description: The ID of the user to get progress for.
          schema:
            type: integer
        - name: quiz_id
          in: query
          description: Filter by specific quiz
          schema:
            type: integer
        - name: days
          in: query
          description: Number of days to analyze
          schema:
            type: integer
            default: 30
            minimum: 1
            maximum: 365
      responses:
        '200':
          description: Retrieved progress data
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      total_sessions:
                        type: integer
                      learn_sessions:
                        type: integer
                      test_sessions:
                        type: integer
                      days_analyzed:
                        type: integer
                      quiz_filter:
                        type: integer
                        nullable: true
                      daily_activity:
                        type: object
                        additionalProperties:
                          type: object
                          properties:
                            learn_sessions:
                              type: integer
                            test_sessions:
                              type: integer
                            average_score:
                              type: number
                              nullable: true
                  message:
                    type: string
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/leaderboard:
    get:
      summary: Get user leaderboard
      tags:
        - users
      description: Get user leaderboard based on performance.
      security:
        - bearerAuth: []
      parameters:
        - name: quiz_id
          in: query
          description: Filter by specific quiz
          schema:
            type: integer
        - name: mode
          in: query
          description: Filter by session mode
          schema:
            type: string
            enum:
              - test
              - learn
            default: test
        - name: limit
          in: query
          description: Number of top users
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Retrieved leaderboard data
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      most_active:
                        type: array
                        items:
                          type: object
                          properties:
                            user_id:
                              type: integer
                            username:
                              type: string
                            session_count:
                              type: integer
                      best_performers:
                        type: array
                        items:
                          type: object
                          properties:
                            user_id:
                              type: integer
                            username:
                              type: string
                            average_score:
                              type: number
                            test_count:
                              type: integer
                      criteria:
                        type: object
                        properties:
                          quiz_id:
                            type: integer
                            nullable: true
                          mode:
                            type: string
                            enum:
                              - test
                              - learn
                          limit:
                            type: integer
                  message:
                    type: string
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    BaseResponse:
      type: object
      properties:
        success:
          type: boolean
          default: true
        message:
          type: string
          nullable: true
        timestamp:
          type: string
          format: date-time
...
    UserBase:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          description: User name
    UserCreate:
      allOf:
        - $ref: '#/components/schemas/UserBase'
    UserUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
    User:
      allOf:
        - $ref: '#/components/schemas/UserBase'
        - type: object
          properties:
            id:
              type: integer
            created_at:
              type: string
              format: date-time
              nullable: true
...
    UserResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/User'
    UsersResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/User'
    UserStatsResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/UserStats'
...
    ErrorDetail:
      type: object
      properties:
        field:
          type: string
          nullable: true
        message:
          type: string
        code:
          type: string
          nullable: true
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          default: false
        error:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            $ref: '#/components/schemas/ErrorDetail'
          nullable: true
        request_id:
          type: string
          nullable: true
        timestamp:
          type: string
          format: date-time
...